feat: Optimize JS resources fetching to load in scope of ActivityStream instead of global  - Meeds-io/MIPs#52 - MEED-1891 (#673)
This change will:
* Change definition of ActivityStream extension for notes to use new
strategy using `load-group` to be loaded only when needed in
ActivityStream
* Move vueCustomElement third party library definition from common
definition to be used only here in Notes
* Change icons to use FA icons instead of deprecated ones
* Delete usage of useless mixins that needs to define one single
property
* Delete useless CSS files
* Delete useless gtmpl files
