+++ notes-webapp/src/main/webapp/WEB-INF/portlet.xml
@@ -51,25 +51,6 @@
       <title>Notes Editor Application</title>
       <keywords>Notes Editor Application</keywords>
     </portlet-info>
-   </portlet>
+  </portlet>
 
-   <portlet>
-    <portlet-name>NotesExtensions</portlet-name>
-    <display-name>Notes Extensions Application</display-name>
-    <portlet-class>org.exoplatform.commons.api.portlet.GenericDispatchedViewPortlet</portlet-class>
-    <init-param>
-      <name>portlet-view-dispatched-file-path</name>
-      <value>/html/notesExtensions.html</value>
-    </init-param>
-    <expiration-cache>-1</expiration-cache>
-    <cache-scope>PUBLIC</cache-scope>
-    <supports>
-    <mime-type>text/html</mime-type>
-    </supports>
-    <supported-locale>en</supported-locale>
-    <portlet-info>
-      <title>Notes Extensions Application</title>
-      <keywords>Notes Extensions Application</keywords>
-    </portlet-info>
-   </portlet>
 </portlet-app>
