--- notes-webapp/src/main/webapp/skin/DefaultSkin/webui/Stylesheet.css
@@ -1,2 +1,0 @@
-@import url('UIWikiRichTextEditor/Stylesheet.css');
-@import url('/eXoSkin/skin/css/wiki/wiki.css');
