--- notes-webapp/src/main/webapp/skin/less/notes/InlineEditing/Style.less
@@ -1,26 +1,0 @@
-.EditSection [class^="wikimodel-freestanding"]{
-  background-position: -40px 0px;
-    &:before {
-        content: "\e791";
-    }
-    margin-top: 12px;
-    .font-icon();
-  &:hover {
-      color: @colorIconPrimary;
-  }
-}
-
-.uiWikiPageContentArea{
-	span.EditSection{
-		display: none;
-	}
-	.section-container .EditSectionHover{
-		background-color: @primaryBackgroundLight;
-    max-width: 100%;
-	}
-	.section-container{
-		padding-top: 5px;
-		padding-bottom: 5px;
-	}
-}
-
