+++ notes-webapp/src/main/webapp/skin/less/notes/Notifications/Style.less
@@ -4,14 +4,14 @@
   border: 1px solid @warningBorder;
   color: #333;
   position: relative;
+  .fa-exclamation-triangle();
   &:before {
-    .font-icon();
+    .fa();
     left: 22px ~'; /** orientation=lt */ ';
     right: 22px ~'; /** orientation=rt */ ';
     top: 9px;
     position: absolute;
     color: @warningColor;
-    content: "\e771";// uiIconWarning
   }
 }
 
@@ -20,13 +20,13 @@
   border: 1px solid @successBorder;
   color: #333;//alway #333
   position: relative;
+  .fa-check-circle();
   &:before {
-    .font-icon();
+    .fa();
     left: 22px ~'; /** orientation=lt */ ';
     right: 22px ~'; /** orientation=rt */ ';
     top: 9px;
     position: absolute;
-    content: "\e770";// uiIconValidate
     color: @successColor;
   }
 }
@@ -36,13 +36,13 @@
   border: 1px solid @infoBorder;
   color: #333;//alway #333
   position: relative;
+  .fa-info-circle();
   &:before {
-    .font-icon();
+    .fas();
     left: 22px ~'; /** orientation=lt */ ';
     right: 22px ~'; /** orientation=rt */ ';
     top: 9px;
     position: absolute;
-    content: "\e767";// uiIconInformation
     color: @infoColor;
   }
 }
@@ -52,13 +52,13 @@
   border: 1px solid @errorBorder;
   color: #333;//alway #333
   position: relative;
+  .fa-exclamation-circle();
   &:before {
-    .font-icon();
+    .fas();
     left: 22px ~'; /** orientation=lt */ ';
     right: 22px ~'; /** orientation=rt */ ';
     top: 9px;
     position: absolute;
-    content: "\e76d";
     color: @errorColor
   }
 }
