--- notes-webapp/src/main/webapp/skin/less/notes/PageInfo/Style.less
@@ -1,57 +1,0 @@
-.uiWikiPageInfo {
-	.uiPageInfoItem {
-		margin: 0px 0px 20px;
-		h6.pageInfoTitle {
-			background: @primaryBackground;
-			border: 1px solid @borderColor;
-			border-bottom: none;
-			line-height: 35px;
-			font-weight: bold;
-			padding: 0 15px;
-			margin: 0;
-		}
-
-		.pageInfoItemBody {
-			background: @baseBackground;
-			border: 1px solid @borderColor;
-			padding: 10px;
-		}
-
-		.uiGrid {
-			margin: 0;
-			table-layout: fixed;
-			word-wrap: break-word;
-		}
-
-		.name {
-			color: @linkColor;
-		}
-
-		.uiAction,
-		.actionCenter {
-			border-top: 1px solid @borderColor;
-			background: @baseBackground;
-			text-align: center;
-			padding: 10px;
-		}
-
-
-	}
-	.uiPageInfoChanges{
-			table{
-				th, td{
-					border-left: none ~'; /** orientation=lt */ ';
-					border-right: 1px solid @borderColor ~'; /** orientation=lt */ ';
-					border-right: none ~'; /** orientation=rt */ ';
-					border-left: 1px solid @borderColor ~'; /** orientation=rt */ ';
-				}
-			}
-		}
-
-
-}
-.uiPageToolBar {
-	.uiDropdownWithIcon {
-		margin-top: 5px;
-	}
-}
