--- notes-webapp/src/main/webapp/skin/less/notes/PopupRelated/Style.less
@@ -1,10 +1,0 @@
-.popupContenRelated{
-	border: solid 1px @borderColor;
-	background: @primaryBackgroundLight;
-	margin: 15px 0 0;
-}
-
-.xUploadPanel .gwt-FileUpload {
-	height: auto;
-	line-height: normal;
-}
