+++ notes-webapp/src/main/webapp/skin/less/notes/Sprites/Style.less
@@ -74,9 +74,3 @@ a:hover [class^="uiIconWiki"],
   color: @colorIconSecondary
 
 }
-
-.uiCompanyNavigationPortlet .uiIconWiki{
-	top: 2px;
-}
-.uiIconWikiRestrictedFile:before {content: "\e6b2";} 
-
