--- notes-webapp/src/main/webapp/skin/less/notes/UINotesMovePageForm/Style.less
@@ -1,41 +1,0 @@
-.uiWikiMovePageForm{
-	h6{
-		margin-top: 0px;
-	}
-	label{
-		font-size: 13px;
-		font-weight: bold;
-	}
-	.selectSpaceForm{
-		label{
-			width: 152px;
-			padding-top: 2px;
-		}
-		.btn-group > .btn:first-child{
-				width: 358px;
-				text-align: start;
-		}
-	}
-	.sideBarContent{
-		.barContent{
-			margin: 10px 0;
-			font-weight: bold;
-		}
-		.sideContent{
-			border: 1px solid @primaryBackgroundHover;
-			background: @primaryBackgroundLight;
-			height: 270px;
-			overflow: auto;
-		}
-	}
-    .uiTreeExplorer {
-        overflow: visible;
-    }
-	.uiWikiLocationContainer{
-		margin: 10px 0 0;
-		label{
-			width: 122px;
-			padding-top: 4px;
-		}
-	}
-}
