--- notes-webapp/src/main/webapp/skin/less/notes/UINotesPageContent/Style.less
@@ -1,113 +1,0 @@
-.uiWikiContentDisplay, .uiWikiPagePreview, .UIWikiRichTextEditor {
-	overflow: auto;
-
-	ul {
-		margin: 10px 18px;
-		padding: 0;
-	}
-	ul li {
-		list-style: disc outside none;
-	}
-	ul li a {
-		text-decoration: none;
-	}
-	ol {
-		margin: 10px 18px;
-		padding: 0;
-	}
-	ol li {
-		list-style: decimal outside none;
-	}
-
-	hr {
-		border-color: @borderColor;
-		border-style: solid;
-		border-width: 1px 0 0;
-		clear: both;
-		height: 0;
-	}
-	dt {
-		font-weight: bold;
-	}
-	dd {
-		background: url("@{images-path}/wiki/Icons/bullet-3.png") no-repeat scroll 0 0.3em transparent;
-		font-style: italic;
-		margin-left: 0 ~'; /** orientation=lt */ ';
-		padding-left: 1em ~'; /** orientation=lt */ ';
-		margin-right: 0 ~'; /** orientation=rt */ ';
-		padding-right: 1em ~'; /** orientation=rt */ ';
-	}
-	blockquote {
-		background: @primaryBackgroundLight;
-		border-left: 2px solid @borderColor ~'; /** orientation=lt */ ';
-		border-right: 2px solid @borderColor ~'; /** orientation=rt */ ';
-		display: block;
-		margin: 0.3ex 1em;
-		padding: 0.3ex 1em;
-		quotes: none;
-	}
-	table {
-		border: 1px solid @borderColor;
-		border-right: none ~'; /** orientation=lt */ ';
-		border-left: none ~'; /** orientation=rt */ ';
-		border-top: none;
-		table-layout: fixed;
-		width: auto;
-	}
-	th {
-		background: @primaryBackground;
-		border-right: 1px solid @borderColor ~'; /** orientation=lt */ ';
-		border-left: 1px solid @borderColor ~'; /** orientation=rt */ ';
-		border-top: 1px solid @borderColor;
-		padding: 5px;
-		text-align: center;
-	}
-	td {
-		border-right: 1px solid @borderColor ~'; /** orientation=lt */ ';
-		border-left: 1px solid @borderColor ~'; /** orientation=rt */ ';
-		border-top: 1px solid @borderColor;
-		padding: 5px;
-	}
-	figure {
-		margin: 0 0 10px;
-		&.image {
-			text-align: center;
-
-			&.image-style-align-left img {
-				float: left ~'; /** orientation=lt */ ';
-				float: right ~'; /** orientation=rt */ ';
-				margin-right: 1.5em ~'; /** orientation=lt */ ';
-				margin-left: 1.5em ~'; /** orientation=rt */ ';
-			}
-
-			&.image-style-align-right img {
-				float: right ~'; /** orientation=lt */ ';
-				margin-left: 1.5em ~'; /** orientation=lt */ ';
-				float: left ~'; /** orientation=rt */ ';
-				margin-right: 1.5em ~'; /** orientation=rt */ ';
-			}
-		}
-	}
-	mark {
-		&.marker-yellow {
-			background-color: #fdfd77;
-		}
-		&.marker-green {
-			background-color: #63f963;
-		}
-		&.marker-pink {
-			background-color: #fc7999;
-		}
-		&.marker-blue {
-			background-color: #72cdfd;
-		}
-		&.pen-red {
-			color: #e91313;
-			background-color: transparent;
-		}
-		&.pen-green {
-			color: #118800;
-			background-color: transparent;
-		}
-	}
-}
