--- notes-webapp/src/main/webapp/skin/less/notes/UINotesPageEditForm/Style.less
@@ -1,148 +1,0 @@
-.uiWikiPageEditForm{
-	padding-bottom: 10px;
-	.actionContainer{
-		white-space: normal;
-		min-height: 30px;
-		.uiSubmitToolBar{
-			margin:0px 0px 0px 5px ~'; /** orientation=lt */ ';
-			margin:0px 5px 0px 0px ~'; /** orientation=rt */ ';
-			float: right ~'; /** orientation=lt */ ';
-			float: left ~'; /** orientation=rt */ ';
-		}
-		.uiWikiPageEditForm_MessageArea  {
-			display: inline-block;
-			vertical-align: middle;
-			font-size: @baseFontSize;
-		}
-		.btn-toolbar {
-			margin:10px 0;
-			.btn-group {
-				margin:0 10px 0px 0 ~'; /** orientation=lt */ ';
-				margin:0 0 0px 10px ~'; /** orientation=rt */ ';
-			}
-		}
-		.uiCheckbox{
-			margin-top: 4px;
-		}
-		.btn-toolbar{
-			.btn-group{
-				.uiIconWikiRickText, .uiIconWatchPage {
-					margin-right: 5px ~'; /** orientation=lt */ ';
-					margin-left: 5px ~'; /** orientation=rt */ ';
-				}
-			}
-		}
-	}
-	.uiEditUpper input{
-		height: 33px;
-		width: 100%;
-		.box-sizing(border-box);
-		margin-bottom: 4px;
-		font-size: 18px;
-	}
-	.uiWikiSidePanelArea{
-		.border-radius (5px 5px 0 0);
-		float: right ~'; /** orientation=lt */ ';
-		float: left ~'; /** orientation=rt */ ';
-		line-height: 16px;
-		margin: 0 0 10px;
-		padding: 0;
-		width: 270px;
-		.barTit{
-			background: @primaryBackground;
-			border: 1px solid @borderColor;
-			.border-radius( 5px 5px 0 0);
-			color: @textColor;
-			font-weight: bold;
-			line-height: 33px;
-			padding-left: 15px ~'; /** orientation=lt */ ';
-			padding-right: 15px ~'; /** orientation=rt */ ';
-			.uiIconClose{
-				margin: 9px 8px 0 0 ~'; /** orientation=lt */ ';
-				margin: 9px 0 0 8px ~'; /** orientation=rt */ ';
-    			text-decoration: none;
-			}
-		}
-		.content{
-			background: @baseBackground;
-			border: 1px solid @borderColor;
-			border-top: none;
-			padding: 12px 8px;
-			.border-radius( 0 0 5px 5px);
-			.box-shadow(0px 1px 3px rgba(0,0,0,0.06));
-			p{
-				margin-bottom: 0px;
-			}
-		}
-	}
-	.uiColsLeftsEditForm{
-		margin-bottom: 22px;
-		margin-right: 290px ~'; /** orientation=lt */ ';
-		margin-left: 290px ~'; /** orientation=rt */ ';
-		padding: 0;
-	}
-	.uiWikiPageContentInputContainer{
-		.commentInput{
-			margin-bottom:24px;
-		}
-		textarea{
-			font-size: @baseFontSize;
-			height: 357px;
-			overflow: auto;
-			padding: 6px;
-			width: 100% !important;
-			.box-sizing(border-box);
-			box-shadow: 0 2px 1px rgba(0, 0, 0, 0.075) inset, 0px 1px 3px rgba(0,0,0,0.06);
-			min-height: 75px;
-		}
-
-		 textarea:focus{
-			box-shadow: 0 1px 1px rgba(0, 0, 0, 0.075) inset, 0 0 5px rgba(29, 75, 159, 0.6);
-		}
-
-	}
-	.uiWikiPageContainer{
-		.form-horizontal {
-			.control-group{
-				.control-label{
-					float: left ~'; /** orientation=lt */ ';
-					float: right ~'; /** orientation=rt */ ';
-					padding: 4px 0 0;
-					margin-right: 10px ~'; /** orientation=lt */ ';
-					margin-left: 10px ~'; /** orientation=rt */ ';
-					text-align: start;
-					width: auto;
-				}
-				.controls{
-					overflow: hidden;
-					display: block;
-					margin-left: 0 ~'; /** orientation=lt */ ';
-					margin-right: 0 ~'; /** orientation=rt */ ';
-				}
-				.controls input{
-					width: 100%;
-				}
-				.publishActivityCheckBoxTitle {
-					margin-left: 10px ~'; /** orientation=lt */ ';
-					margin-right: 10px ~'; /** orientation=rt */ ';
-				}
-			}
-			.uiCheckbox{
-				margin-top: 3px;
-			}
-		}
-	}
-
-	.uiWikiTemplateDescriptionContainer{
-		margin-top: 10px;
-		input{
-			margin: 0px;
-			width: 100%;
-			.box-sizing(border-box);
-			height: 28px;
-		}
-	}
-	.uiWikiNotificationContainer a:not(.btn) {
-		color: @primaryColor;
-	}
-}
