--- notes-webapp/src/main/webapp/skin/less/notes/UINotesPageVersionsCompare/Style.less
@@ -1,79 +1,0 @@
-.uiWikiPageVersionsCompare .comparedForm {
-	border-bottom: 1px solid @primaryBackgroundHover;
-	padding-bottom: 20px;
-}
-
-.uiWikiPageVersionsCompare .comparedForm pre {
-	background-color: @primaryBackgroundLight;
-    border: 1px solid @primaryBackgroundHover;
-    .border-radius(2px);
-    display: block;
-    white-space: normal;
-    padding: 18px;
-    line-height: 25px;
-    #font > #family > .sans-serif;
-
-    > b {
-    	&.admin {
-    		color: @linkColor;
-    	}
-    }
-
-    > span {
-    	&.dateChange {
-    		display: block;
-    	}
-
-    	&.lineRemoved {
-    		background-color: @wikiPinkLighter;
-		    text-decoration: line-through;
-            color: #333;
-    	}
-
-    	&.lineAdded {
-    		background-color: @wikiGreenLighter;
-            color: #333;
-    	}
-    }
-
-    > a.changes {
-		color: @linkColor;
-	}
-}
-
-.comparedWith {
-	font-weight: bold;
-    padding: 30px 28px 0;
-}
-
-.uiWikiPageVersionsCompare .changedForm {
-	margin-top: 20px;
-	> button {
-		margin: 0 0 10px;
-	}
-
-	> .diff > .diffmodifiedline {
-		> span {
-			&.diffaddword {
-				background-color: @wikiGreenLighter;
-			}
-			&.diffremoveword {
-				background-color: @wikiPinkLighter;
-			}
-		}
-		> pre {
-			border: none;
-			border-radius: 0;
-			margin: 0;
-			padding: 0;
-			white-space: normal;
-			&.diffaddword {
-				background-color: @wikiGreenLighter;
-			}
-
-			&.diffremoveword {
-				background-color: @wikiPinkLighter;
-			}
-		}
-	}
-}
