--- notes-webapp/src/main/webapp/skin/less/notes/UINotesPermalinkForm/Style.less
@@ -1,10 +1,0 @@
-.uiWikiPermalinkForm {
-	.permalinkWapperInput {
-		.input-xxxlarge {
-			width: 100%;
-		}		 
-	}
-	.permalLinkShare {
-		margin-bottom: 10px;
-	}
-}
