+++ notes-webapp/src/main/webapp/skin/less/notes/UIWikiRichTextEditor/Stylesheet.less
@@ -1,52 +1,3 @@
-.UIWikiRichTextEditor {
-  height: 500px;
-  background-color: #FFFFFF;
-  border: 1px solid var(--ck-color-base-border);
-  border-radius: var(--ck-border-radius);
-  margin-left: 1px;
-  margin-right: 1px;
-}
-
-.UIWikiRichTextEditor .wiki-children-pages.ck-widget, .wiki-include-page.ck-widget {
-  position: relative;
-}
-
-/* workaround for children-pages and include-page widgets selection */
-.UIWikiRichTextEditor .wiki-children-pages.ck-widget:before, .wiki-include-page.ck-widget:before {
-  content: "";
-  top: 0;
-  left: 0; /** orientation=lt */
-  right: 0; /** orientation=rt */
-  width: 100%;
-  height: 100%;
-}
-
-#UIWikiRichTextArea_TextArea {
-  display: none;
-}
-
-.UIWikiEditorToolbar {
-  position: relative;
-  z-index: 1;
-  margin-top: 10px;
-  border-bottom: 0;
-  box-shadow: 0 0 5px hsla( 0,0%,0%,.2 );
-}
-
-/* custom icon on ckeditor buttons */
-.UIWikiEditorToolbar .ck.ck-button[class*="uiIcon"] {
-  font-family: "PLF-FONT-ICONS";
-}
-
-.UIWikiEditorToolbar .document-editor__toolbar .ck-toolbar {
-  border: 0;
-  border-radius: 0;
-}
-
-.UIWikiRichTextEditor.ck-editor__editable {
-  min-height: 380px;
-}
-
 .ck-toolbar .ck-heading-dropdown .ck-list__item {
   padding: 2px 4px;
 }
