+++ notes-webapp/src/main/webapp/skin/less/notes/notes-mixins.less
@@ -9,7 +9,7 @@
 		font-weight: bold;
 		padding: 0 15px;
 		margin: 0;
-		.box-shadow(1px 0px 2px rgba(0,0,0,0.2));
+		box-shadow: 1px 0px 2px rgba(0,0,0,0.2);
 	}
 }
