+++ notes-webapp/src/main/webapp/skin/less/notes/notes.less
@@ -1,42 +1,14 @@
 @import "../variables.less";
 @import "../mixins.less";
+
 @import "notes-variables.less";
 @import "notes-mixins.less";
-@import "UINotesPermalinkForm/Style.less";
+
 @import "Notifications/Style.less";
-@import "UINotesPageVersionsCompare/Style.less";
-@import "UINotesPageEditForm/Style.less";
-@import "InlineEditing/Style.less";
-@import "PageInfo/Style.less";
-@import "PopupRelated/Style.less";
-@import "UINotesAdvanceSearchResult/Style.less";
-@import "UINotesMovePageForm/Style.less";
 @import "Sprites/Style.less";
-@import "UIViewContentDisplay/Style.less";
-@import "UINotesPageContent/Style.less";
 @import "UIWikiRichTextEditor/Stylesheet.less";
-@import "../social/skin/VersionHistoryDrawer/style.less";
-
-.UIWikiToolBar {
-  float: right ~'; /** orientation=lt */ ';
-  float: left ~'; /** orientation=rt */ ';
-  width: 100px;
-}
-
-.uiWikiSearchBox {
-
-  .uiSearchInput input[type="text"] {
-    //background: none;
-  }
-
-  &.pull-right > .dropdown-menu {
-    right: auto ~'; /** orientation=lt */ ';
-    left: auto ~'; /** orientation=rt */ ';
-  }
-}
 
 /************ New css feature notes ****************/
-
 .customPluginsDrawer {
   z-index: 99999 !important;
 
@@ -85,8 +57,6 @@
     z-index: 99999 !important;
 
     .uiArrowBAckIcon:before {
-      content: "\e913";
-      font-family: "PLF-FONT-ICONS" !important;
       font-style: normal;
       font-size: 22px;
     }
@@ -335,13 +305,8 @@
 
       .notes-treeview {
         .uiTreeviewIcon {
-          font-family: "PLF-FONT-ICONS" !important;
           font-style: normal;
           font-size: 20px;
-
-          &:before {
-            content: "\e698";
-          }
         }
       }
     }
@@ -425,30 +390,11 @@
 
   .drawerContent {
     .search {
-      margin: 0!important;
       padding: 0!important;
+      margin-bottom: 1px;
+
       .mdi-close {
         font-size: 17px;
-      }
-    }
-    .filter {
-      .btn-group, .btn.dropdown-toggle {
-        width: inherit;
-      }
-      .btn.dropdown-toggle {
-        border: 1px solid @greyColorDefault !important;
-      }
-
-      ul {
-        width: auto !important;
-
-        a {
-          color: @darkGrey !important;
-
-          &:hover {
-            color: white!important;
-          }
-        }
       }
     }
     .checkbox {
@@ -561,12 +507,12 @@
 .wikiBox {
   background: @baseBackground;
   border: 1px solid @borderColor;
-  .border-radius(5px);
-  .box-shadow(0px 1px 2px rgba(0, 0, 0, 0.08));
+  border-radius: 5px;
+  box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.08);
 
   .titleWikiBox {
-    .border-radius(5px 5px 0px 0px);
-    .box-shadow(0px 1px 2px rgba(13, 13, 13, 0.08));
+    border-radius: 5px 5px 0px 0px;
+    box-shadow: 0px 1px 2px rgba(13, 13, 13, 0.08);
     background: @primaryBackground;
     border-bottom: 1px solid @borderColor;
     margin: 0px;
@@ -688,7 +634,7 @@
 }
 
 .fileHidden {
-  .opacity(0);
+  opacity: 0;
   position: relative;
   z-index: 2;
   outline: none;
@@ -749,9 +695,11 @@
     }
 
     .resizeButton {
-      #gradient .vertical(@baseBackground, @primaryBackground);
-      .border-radius(@borderRadiusSmall 0px 0px @borderRadiusSmall);
-      .size(24px, 14px);
+      background-image: linear-gradient(to bottom, @baseBackground, @primaryBackground);
+      background-repeat: repeat-x;
+      border-radius: @borderRadiusSmall 0px 0px @borderRadiusSmall;
+      height: 24px;
+      width: 14px;
       float: right ~'; /** orientation=lt */ ';
       float: left ~'; /** orientation=rt */ ';
       margin: auto;
@@ -802,7 +750,7 @@
         border: 1px solid @borderColor;
         border-left: none ~'; /** orientation=lt */ ';
         border-right: none ~'; /** orientation=rt */ ';
-        .border-radius(0px @borderRadiusSmall @borderRadiusSmall 0px);
+        border-radius: 0px @borderRadiusSmall @borderRadiusSmall 0px;
         float: left ~'; /** orientation=lt */ ';
         float: right ~'; /** orientation=rt */ ';
       }
@@ -1017,18 +965,6 @@
     }
   }
 
-  .uiWikiMovePageForm, .uiWikiSelectPageForm {
-    .labelForm {
-      font-weight: bold;
-      padding: 4px 0px;
-    }
-
-    .uiInputInfo {
-      font-weight: bold;
-      padding: 4px 0px 10px;
-    }
-  }
-
   .UISelectPageForm .SpaceSwitcher {
     padding: 15px 15px 0px;
   }
@@ -1138,16 +1074,6 @@
     //  background-image: none;
   }
 }
-
-/******* wiki search card ************/
-.wikiSearchCard {
-  .uiIconWiki:before {
-    font-family: "PLF-FONT-ICONS";
-    content: "\e695";
-    font-size: unset;
-  }
-}
-
 
 //responsive wiki
 @media (min-width: 769px) and (max-width: 1024px) {
