+++ notes-webapp/src/main/webapp/vue-app/notes-favorite-drawer-extension/extensions.js
@@ -1,0 +1,6 @@
+export function initNotesExtensions() {
+  extensionRegistry.registerComponent('favorite-notes', 'favorite-drawer-item', {
+    id: 'notes',
+    vueComponent: Vue.options.components['notes-favorite-item'],
+  });
+}
