+++ notes-webapp/src/main/webapp/vue-app/notes-favorite-drawer-extension/initComponents.js
