+++ notes-webapp/webpack.prod.js
@@ -14,7 +14,8 @@ const config = {
     notes: './src/main/webapp/vue-app/notes/main.js',
     notesEditor: './src/main/webapp/vue-app/notes-editor/main.js',
     notesSwitch: './src/main/webapp/vue-app/notes-switch/main.js',
-    notesExtension: './src/main/webapp/vue-app/notes-extensions/main.js',
+    notesActivityStreamExtension: './src/main/webapp/vue-app/notes-activity-stream-extension/main.js',
+    notesFavoriteDrawerExtension: './src/main/webapp/vue-app/notes-favorite-drawer-extension/main.js',
     engagementCenterExtensions: './src/main/webapp/vue-app/engagementCenterExtensions/extensions.js'
   },
   output: {
